COM-2927: move functions to helper
